Reconfiguring whole project as a package named "phidl"
