+++ phidl/__init__.py
@@ -1,0 +1,2 @@
+from phidl.device_layout import Device, SubDevice, Port
+from phidl.device_layout import quickplot, inset
